Refactoring all the code to look more

+++ src/screens/TransactionDetailScreen.js
@@ -13,28 +13,38 @@ import base64js from 'base64-js'
 import crypto from 'crypto-js'
 import sha256 from 'crypto-js/sha256';
 import { get, sortBy } from 'lodash'
-import ErrorMessage from './../components/ErrorMessage'
-import SecurityForm from './../components/SecurityForm'
-import DetailTabs from './../components/DetailTabs'
-import { Screen, ContainerFlex, Container, SelectSecret, Header, Title, TitleWrapper, LoadButtonWrapper, LoadButton } from './../components/utils'
+import ErrorMessage from './../components/UI/ErrorMessage'
+import SecurityForm from './../components/UI/SecurityForm'
+import DetailTabs from './../components/DetailTab/DetailTabs'
 import { decodeFromXdr, signXdr } from './../utils/xdrUtils';
+import { generateKeypair } from './../utils/bipUtil';
 import PouchDB from 'pouchdb-react-native'
 import SQLite from 'react-native-sqlite-2'
 import SQLiteAdapterFactory from 'pouchdb-adapter-react-native-sqlite'
+import {
+	Screen,
+	ContainerFlex, 
+	Container, 
+	Header, 
+	Title, 
+	TitleWrapper, 
+	LoadButtonWrapper, 
+	LoadButton
+} from './styled';
+
 
 const SQLiteAdapter = SQLiteAdapterFactory(SQLite)
 PouchDB.plugin(SQLiteAdapter)
 const db = new PouchDB('Secrets', { adapter: 'react-native-sqlite' })
 const db2 = new PouchDB('Transactions', { adapter: 'react-native-sqlite' })
-import { generateKeypair } from './../utils/bipUtil';
 
 @inject('appStore') @observer
 class TransactionDetail extends Component {
 
 	static navigationOptions = ({ navigation }) => {
-    const params = navigation.state.params || {};
+		const params = navigation.state.params || {};
 
-    return {
+		return {
 			header: (
 				<SafeAreaView style={{ backgroundColor: '#2e3666' }}>
 					<Header>
@@ -42,10 +52,10 @@ class TransactionDetail extends Component {
 							<Title>Transaction Detail</Title>
 						</TitleWrapper>
 						<LoadButtonWrapper>
-							<LoadButton onPress={()=> navigation.goBack()}>
+							<LoadButton onPress={() => navigation.goBack()}>
 								<Icon name="x-circle" color="white" size={32} />
 							</LoadButton>
-						</LoadButtonWrapper>						
+						</LoadButtonWrapper>
 					</Header>
 				</SafeAreaView>
 			)
@@ -53,29 +63,29 @@ class TransactionDetail extends Component {
 	};
 
 	state = {
-    tabView: {
-      index: 0,
-      routes: [
-        { key: 'display', title: 'Operation' },
-        { key: 'envelop', title: 'Envelope' },
-        { key: 'signed', title: 'Signed' }
-      ]
+		tabView: {
+			index: 0,
+			routes: [
+				{ key: 'display', title: 'Operation' },
+				{ key: 'envelop', title: 'Envelope' },
+				{ key: 'signed', title: 'Signed' }
+			]
 		},
 		secrets: [],
 		options: [],
-    showSecurityForm: false
+		showSecurityForm: false
 	}
 
-  componentDidMount() {
+	componentDidMount() {
 		this.loadData();
 	}
-	
+
 	loadData = () => {
 		try {
 			let self = this;
 			db.allDocs({
 				include_docs: true
-			}).then((res)=> {
+			}).then((res) => {
 				const options = [];
 				const secrets = res.rows;
 				secrets.forEach(el => options.push(el.doc.alias));
@@ -85,7 +95,7 @@ class TransactionDetail extends Component {
 			alert(error.message)
 		}
 	}
-	
+
 	copyToClipboard = () => {
 		const { appStore, navigation } = this.props
 		const tx = appStore.get('currentTransaction')
@@ -93,22 +103,22 @@ class TransactionDetail extends Component {
 		alert('The signed xdr was copied to the clipboard.');
 	}
 
-  handleTabIndexChange = index => {
-    this.setState({
-      tabView: Object.assign({}, this.state.tabView, {
-        index
-      })
-    })
-  }
+	handleTabIndexChange = index => {
+		this.setState({
+			tabView: Object.assign({}, this.state.tabView, {
+				index
+			})
+		})
+	}
 
-  signTransaction = () => {
-    this.authTransaction();
-  }
+	signTransaction = () => {
+		this.authTransaction();
+	}
 
-  authTransaction = () => {
+	authTransaction = () => {
 		const { appStore, navigation } = this.props
 		const seed = appStore.get('seed')
-		
+
 		const { secrets } = this.state;
 		if (!secrets || secrets.length === 0) {
 			Alert.alert(
@@ -126,25 +136,25 @@ class TransactionDetail extends Component {
 		}
 	}
 
-  rejectTransaction = () => {
+	rejectTransaction = () => {
 		const { appStore, navigation } = this.props
 		const currentTransaction = appStore.get('currentTransaction')
 		try {
-      db2.put({
+			db2.put({
 				_id: currentTransaction._id,
 				...currentTransaction,
 				status: 'REJECTED'
 			});
 			navigation.goBack()
-	    setTimeout(()=> {
+			setTimeout(() => {
 				appStore.set('currentTransaction', undefined)
 			}, 1000)
 		} catch (error) {
 			alert(error.message)
 		}
-  }
+	}
 
-  deleteTransaction = async (currentTransaction) => {
+	deleteTransaction = async (currentTransaction) => {
 		const { appStore, navigation } = this.props
 		let ctx;
 		try {
@@ -155,37 +165,37 @@ class TransactionDetail extends Component {
 			}
 			const res = await db2.remove(ctx);
 			navigation.goBack()
-			setTimeout(()=> {
+			setTimeout(() => {
 				appStore.set('currentTransaction', undefined)
 			}, 1000)
 		} catch (error) {
 			alert(error.message);
 		}
 	}
-	
-	showConfirmDelete = tx => {
-    Alert.alert(
-      `Are you sure you want delete this?`,
-      `${tx.memo}`,
-      [
-        { text: 'Cancel', onPress: () => {}, style: 'cancel' },
-        {
-          text: 'Confirm',
-          onPress: () => this.deleteTransaction(tx)
-        }
-      ],
-      { cancelable: true }
-    )
-  }
 
-  submitSignature = index => {
-    const { secrets } = this.state;
-		const secret = secrets[index]
-    this.showConfirmSignatureAlert(secret)
+	showConfirmDelete = tx => {
+		Alert.alert(
+			`Are you sure you want delete this?`,
+			`${tx.memo}`,
+			[
+				{ text: 'Cancel', onPress: () => { }, style: 'cancel' },
+				{
+					text: 'Confirm',
+					onPress: () => this.deleteTransaction(tx)
+				}
+			],
+			{ cancelable: true }
+		)
 	}
-	
-  confirmSignTransaction = secret => {
-    const { appStore, navigation } = this.props
+
+	submitSignature = index => {
+		const { secrets } = this.state;
+		const secret = secrets[index]
+		this.showConfirmSignatureAlert(secret)
+	}
+
+	confirmSignTransaction = secret => {
+		const { appStore, navigation } = this.props
 		try {
 			const seed = appStore.get('seed');
 			const keypair = generateKeypair(seed, secret.vn);
@@ -193,9 +203,9 @@ class TransactionDetail extends Component {
 			const sk = keypair.secret();
 			const ctx = appStore.get('currentTransaction')
 			const data = {
-					type: 'sign',					
-					sk,
-					...ctx,
+				type: 'sign',
+				sk,
+				...ctx,
 			}
 
 			const signedTx = signXdr(data)
@@ -205,62 +215,62 @@ class TransactionDetail extends Component {
 			alert(error.message)
 		}
 	}
-	
+
 	saveCurrentTransaction = data => {
 		const { appStore } = this.props
 		const currentTransaction = appStore.get('currentTransaction')
-    if (data) {
-      if (data.type === 'error') {
-        //console.warn('Error: ', data);
-        this.saveTransaction({
-          xdr: data.xdr,
-          createdAt: new Date().toISOString(),
-          type: 'error',
-          message: data.message,
-          status: 'ERROR'
-        })
-      } else if (data.type === 'sign') {
-        this.saveTransaction({
-          ...currentTransaction,
-          ...data,
+		if (data) {
+			if (data.type === 'error') {
+				//console.warn('Error: ', data);
+				this.saveTransaction({
+					xdr: data.xdr,
+					createdAt: new Date().toISOString(),
+					type: 'error',
+					message: data.message,
+					status: 'ERROR'
+				})
+			} else if (data.type === 'sign') {
+				this.saveTransaction({
+					...currentTransaction,
+					...data,
 					status: 'SIGNED',
 					createdAt: new Date().toISOString()
-        })
-      } else {
+				})
+			} else {
 				const tx = parseEnvelopeTree(data.tx)
-        this.saveTransaction({
+				this.saveTransaction({
 					...tx,
-          type: data.type,
+					type: data.type,
 					xdr: data.xdr,
-          createdAt: new Date().toISOString(),
-          status: 'CREATED'
-        })
-      }
-    } else {
-      console.warn('Data not found!');
-    }
+					createdAt: new Date().toISOString(),
+					status: 'CREATED'
+				})
+			}
+		} else {
+			console.warn('Data not found!');
+		}
 	}
-	
+
 	saveTransaction = async tx => {
-    const { appStore } = this.props
+		const { appStore } = this.props
 		try {
-      db2.put({
+			db2.put({
 				_id: uuid(),
 				...tx
 			});
 		} catch (error) {
 			alert(error.message)
 		}
-    appStore.set('currentXdr', undefined)
-  }
+		appStore.set('currentXdr', undefined)
+	}
 
-  showConfirmSignatureAlert = secret => {
+	showConfirmSignatureAlert = secret => {
 		if (secret && secret.doc) {
 			Alert.alert(
 				`${secret.doc.alias}`,
 				`${secret.doc.pk}`,
 				[
-					{ text: 'Cancel', onPress: () => {}, style: 'cancel' },
+					{ text: 'Cancel', onPress: () => { }, style: 'cancel' },
 					{
 						text: 'Confirm',
 						onPress: () => this.confirmSignTransaction(secret.doc)
@@ -271,47 +281,47 @@ class TransactionDetail extends Component {
 		} else {
 			this.actionSheet.hide(0);
 		}
-  }
+	}
 
-  render() {
-    const { appStore, toggleModal } = this.props
+	render() {
+		const { appStore, toggleModal } = this.props
 		const { showSecurityForm, options } = this.state
 		const currentTransaction = appStore.get('currentTransaction');
 		const securityFormError = appStore.get('securityFormError')
 
-    if (currentTransaction && currentTransaction.type === 'error') {
-      return (
-        <Container>
-          <ErrorMessage tx={currentTransaction} />
-          <Button
-            ref={ref => (this.deleteTransactionButton = ref)}
-            foregroundColor={'white'}
-            backgroundColor={'#ff3b30'}
-            errorColor={'#ff3b30'}
-            errorIconColor={'white'}
-            successIconColor={'white'}
-            onPress={this.deleteTransaction}
-            successIconName="check"
-            label="Delete"
-            maxWidth={100}
-            style={{ marginLeft: 16, borderWidth: 0, alignSelf: 'center' }}
-          />
-        </Container>
-      )
-    }
+		if (currentTransaction && currentTransaction.type === 'error') {
+			return (
+				<Container>
+					<ErrorMessage tx={currentTransaction} />
+					<Button
+						ref={ref => (this.deleteTransactionButton = ref)}
+						foregroundColor={'white'}
+						backgroundColor={'#ff3b30'}
+						errorColor={'#ff3b30'}
+						errorIconColor={'white'}
+						successIconColor={'white'}
+						onPress={this.deleteTransaction}
+						successIconName="check"
+						label="Delete"
+						maxWidth={100}
+						style={{ marginLeft: 16, borderWidth: 0, alignSelf: 'center' }}
+					/>
+				</Container>
+			)
+		}
 
 		return (
 			<ContainerFlex style={{ backgroundColor: '#d5eef7' }}>
-						{!showSecurityForm && (
-							<DetailTabs
-								currentTransaction={currentTransaction}
-								copyToClipboard={this.copyToClipboard}
-								showConfirmDelete={this.showConfirmDelete} 
-								rejectTransaction={this.rejectTransaction}
-								signTransaction={this.signTransaction}
-							/>
-						)}
-						{/* {showSecurityForm && (
+				{!showSecurityForm && (
+					<DetailTabs
+						currentTransaction={currentTransaction}
+						copyToClipboard={this.copyToClipboard}
+						showConfirmDelete={this.showConfirmDelete}
+						rejectTransaction={this.rejectTransaction}
+						signTransaction={this.signTransaction}
+					/>
+				)}
+				{/* {showSecurityForm && (
 							<SecurityForm
 								error={securityFormError}
 								submit={this.authTransaction}
@@ -319,19 +329,19 @@ class TransactionDetail extends Component {
 								closeAfterSubmit={false}
 							/>
 						)} */}
-						{(options && options.length > 0) && (
-							<ActionSheet
-								ref={o => (this.actionSheet = o)}
-								title={'Select a Secret'}
-								options={options}
-								cancelButtonIndex={1}
-								destructiveButtonIndex={2}
-								onPress={this.submitSignature}
-							/>
-						)}
+				{(options && options.length > 0) && (
+					<ActionSheet
+						ref={o => (this.actionSheet = o)}
+						title={'Select a Secret'}
+						options={options}
+						cancelButtonIndex={1}
+						destructiveButtonIndex={2}
+						onPress={this.submitSignature}
+					/>
+				)}
 			</ContainerFlex>
 		)
-  }
+	}
 }
 
 export default TransactionDetail
